feat(auth): implement user authentication flow with helper functions and UI components
- Extracts authentication logic into separate helper functions and type definitions
- Splits profile page components into dedicated tab components for better maintainability
- Removes unused imports and simplifies toast notification positioning
- Improves code structure by separating concerns and reducing component complexity
