+++ learn_nextjs/src/app/auth-provider.types.ts
@@ -1,0 +1,12 @@
+import type { User } from '@/common/types/api/resources'
+
+export type AuthState =
+  | { status: 'loading' }
+  | { status: 'authenticated'; user: User }
+  | { status: 'unauthenticated' }
+
+export interface AuthContextValue {
+  state: AuthState
+  onLoginSuccess: (user: User) => void
+  onLogoutSuccess: () => void
+}
