+++ learn_nextjs/src/app/me/components/index.ts
@@ -1,0 +1,2 @@
+export * from './TabItems'
+export * from './TabProfile'
